requirements.txt
alembic==1.6.5
aniso8601==9.0.1
click==8.0.1
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.13.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
